Return value should be a long as this is an item_id

+++ Source/Podio .NET/Services/ItemService.cs
@@ -357,7 +357,7 @@ namespace PodioAPI.Services
         ///     generated
         /// </param>
         /// <returns>The id of the cloned item</returns>
-        public async Task<int> CloneItem(long itemId, bool silent = false)
+        public async Task<long> CloneItem(long itemId, bool silent = false)
         {
             string url = string.Format("/item/{0}/clone", itemId);
             url = Utility.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent));
